==> 06-normatividad.Rmd <==
# Normatividad

Una de las mediciones fundamentales en la evaluación del impacto de los programas PAES y PEAMA es hacer una validación de la normatividad vigente y aplicable para la cohorte 2011-01, esto debido a que la normatividad refleja que las zonas de impacto de los programas sean las realmente escaladas, que se pertenezca a las poblaciones a quien va realmente dirigida y que se les esten dando las oportunidades que estan reglamentadas por ley y que aportan a la inclusión educativa en la Universidad Nacional de Colombia:

```{r,eval=TRUE,echo=FALSE,results='hide',warning=FALSE,error=FALSE,message=FALSE}
library("tidyverse")
load("NORMAS.RData")
load("CUPOS.RData")
load("CUPOS_PEAMA.RData")
load("P2011_1.RData")
load("ADMITIDOS_2.RData")
PAES <- P2011_1 %>% filter(TIPO_INS=="PAES",ADMITIDO=="Sí")
load("ADMITIDOS_PEAMA.RData")
PEAMA <- ADMITIDOS_PEAMA
PEAMA_2 <- P2011_1 %>% filter(TIPO_INS=="PEAMA",ADMITIDO=="Sí")
PUNTAJE <- P2011_1 %>% filter(TIPO_INS=="Regular",ADMITIDO=="Sí") %>% mutate(P=min(PTOTAL,na.rm = TRUE)) %>%  distinct(P)
rm(P2011_1);gc()

```

```{r,echo=FALSE}
knitr::kable(
  NORMAS, booktabs = TRUE,
  caption = 'Normatividad vigente y aplicable para la cohorte de estudio'
)
```

## Validación de la normatividad vigente y aplicable a los PAES

### Puntaje en examen de admisión

Como dice el **acuerdo 22 de 1986 del CSU en el articulo dos** y el **acuerdo 18 de 1999 del CSU en el articulo dos**  para las _comunidades indigenas_ y el **acuerdo 93 de 1989 del CSU en el articulo dos** para los _mejores bachilleres de municipios pobres_, los estudiantes admitidos por estas dos modalidades deben tener un puntaje de admisión mayor o igual al último admitido de manera regular, además como dice el**acuerdo 30 de 1990 del CSU en el articulo dos** para los _mejores bacilleres_ y el **acuerdo 13 de 2009 del CSU en el articulo dos** para las _poblaciones afrocolombianas_ son admitidos en los mismos términos que los estudiantes de admisión regular, luego para ellos también aplica tener el puntaje de admisión es mayor o igual al último admitido de manera regular. 

Se observa, como una línea roja, el menor puntaje (451.2525) obtenido por los admitidos de manera regular y es claro que en todos los programas PAES se satisface la norma del menor puntaje de admisión posible para ellos.

```{r,echo=FALSE}
df <- PAES %>% select(PAES,PTOTAL) %>% filter(is.na(PTOTAL)==FALSE)

color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df, aes(PAES,PTOTAL, fill=PAES)) +
  geom_boxplot(show.legend = FALSE) +
  scale_fill_manual(values = color) +
  xlab("Programa PAES") +
  ylab("Puntaje en el examen de admisión") + 
  theme_classic() + 
  ggtitle(label = "Puntaje total de admisión de los admitidos PAES", subtitle =  "(Cohorte 2011-01)") +
  geom_hline(yintercept = PUNTAJE$P,color="red") + 
  coord_flip()
```


### Razón de admitidos PAES por programa vs. la cantidad de cupos ofertados por el programa curricular

Como dice el **acuerdo 22 de 1986 del CSU en el articulo uno** para las _comunidades indigenas_ y el **acuerdo 93 de 1989 del CSU en el articulo uno** para los _mejores bachilleres de municipios pobres_ y el **acuerdo 13 de 2009 del CSU en el articulo tres** para las _poblaciones afrocolombianas_ los cuales establecen que se deben destinar por lo menos el $2\%$ de los cupos por carrera a cada uno de los tres programas PAES anteriormente mencionados.

Con el objetivo de validar las normas anteriormente mencionadas se observa a continuación una comparación del número de admitidos de cada uno de los programas PAES por carrera contra la cantidad de cupos.

```{r,echo=FALSE, warning=FALSE, message=FALSE}
ADMITIDOS_3 <- ADMITIDOS_0 %>% 
  filter(ASP_RAYA_TIPO %in% c("Regular","Afrocolombianos","MBMP","IND") ) %>%
  group_by(ADM_SEDE_NOMBRE, ADM_CARRERA_NOMBRE,ASP_RAYA_TIPO) %>%
  count() %>%
  spread(ASP_RAYA_TIPO,n) %>% 
  left_join(CUPOS %>% select(ADM_CARRERA_NOMBRE,CUPO_RESOLUCION))  %>% 
  mutate("Regular"=Regular*100/CUPO_RESOLUCION,
         "Población afrocolombiana"=Afrocolombianos*100/CUPO_RESOLUCION, 
         "Comunidades indígenas"=IND*100/CUPO_RESOLUCION,
         "Mejores bachilleres de municipios pobres"=MBMP*100/CUPO_RESOLUCION) %>% gather("PROGRAMA","PORCENTAJE",`Población afrocolombiana`,`Comunidades indígenas`,`Mejores bachilleres de municipios pobres`)

```


Primero para el caso de la sede Bogotá se tiene que la razón de admitidos PAES comparada con el número de cupos ofertados por programa curricular es en todos los casos igual o superior al 2%. En los casos de la sede Medellín y la sede Manizales se observa que en la mayoría de los casos la razón de admitidos PAES por programa curricular contra la cantidad de cupos disponibles está en 2% o es superior, en unos pocos casos es inferior al 2% pero recordemos que la norma es sensible a un redondeo de la cantidad de cupos que les corresponden a los programas, también se puede dar el caso en que la demanda de un programa es inferior la cantidad de cupos reservados a los PAES. Por otra parte, para la sede Palmira, todos los programas tienen una razón superior al 2%. 

```{r,echo=FALSE, warning=FALSE}
df <- ADMITIDOS_3 %>% filter(ADM_SEDE_NOMBRE=="Bogotá")
color <- c("#1E90FF", "#6E8B3D", "#B23AEE")
ggplot(df, aes(color=PROGRAMA))+ 
  geom_line(aes(ADM_CARRERA_NOMBRE, PORCENTAJE, group = PROGRAMA ), show.legend = FALSE) +
  geom_area(aes(ADM_CARRERA_NOMBRE, PORCENTAJE, group = PROGRAMA ), show.legend = FALSE, alpha=0.2) +
  geom_point(aes(ADM_CARRERA_NOMBRE, PORCENTAJE), show.legend = FALSE) + 
  coord_flip() + 
  labs(fill="Programa PAES") +
  ylab("Porcentaje de admitidos PAES") +
  xlab("Carreras") +
  ggtitle(label = "Admitidos PAES a la sede Bogotá por carrera", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) +
  theme_classic() + 
  geom_hline(yintercept = 2, color="black") +
  facet_wrap(~PROGRAMA) 
```


```{r,echo=FALSE, warning=FALSE}
df <- ADMITIDOS_3 %>% filter(ADM_SEDE_NOMBRE=="Medellín")
color <- c("#1E90FF", "#6E8B3D", "#B23AEE")
ggplot(df, aes(color=PROGRAMA))+ 
  geom_line(aes(ADM_CARRERA_NOMBRE, PORCENTAJE, group = PROGRAMA ), show.legend = FALSE) +
  geom_area(aes(ADM_CARRERA_NOMBRE, PORCENTAJE, group = PROGRAMA ), show.legend = FALSE, alpha=0.2) +
  geom_point(aes(ADM_CARRERA_NOMBRE, PORCENTAJE), show.legend = FALSE) + 
  coord_flip() + 
  labs(fill="Programa PAES") +
  ylab("Porcentaje de admitidos PAES") +
  xlab("Carreras") +
  ggtitle(label = "Admitidos PAES a la sede Medellín por carrera", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) +
  theme_classic() + 
  geom_hline(yintercept = 2, color="black") +
  facet_wrap(~PROGRAMA) 
```

```{r,echo=FALSE, warning=FALSE}
df <- ADMITIDOS_3 %>% filter(ADM_SEDE_NOMBRE=="Manizales")
color <- c("#1E90FF", "#6E8B3D", "#B23AEE")
ggplot(df, aes(color=PROGRAMA))+ 
  geom_line(aes(ADM_CARRERA_NOMBRE, PORCENTAJE, group = PROGRAMA ), show.legend = FALSE) +
  geom_area(aes(ADM_CARRERA_NOMBRE, PORCENTAJE, group = PROGRAMA ), show.legend = FALSE, alpha=0.2) +
  geom_point(aes(ADM_CARRERA_NOMBRE, PORCENTAJE), show.legend = FALSE) + 
  coord_flip() + 
  labs(fill="Programa PAES") +
  ylab("Porcentaje de admitidos PAES") +
  xlab("Carreras") +
  ggtitle(label = "Admitidos PAES a la sede Manizales por carrera", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) +
  theme_classic() + 
  geom_hline(yintercept = 2, color="black") +
  facet_wrap(~PROGRAMA)
```

```{r,echo=FALSE, warning=FALSE}
df <- ADMITIDOS_3 %>% filter(ADM_SEDE_NOMBRE=="Palmira")
color <- c("#1E90FF", "#6E8B3D", "#B23AEE")
ggplot(df, aes(color=PROGRAMA))+ 
  geom_line(aes(ADM_CARRERA_NOMBRE, PORCENTAJE, group = PROGRAMA ), show.legend = FALSE) +
  geom_area(aes(ADM_CARRERA_NOMBRE, PORCENTAJE, group = PROGRAMA ), show.legend = FALSE, alpha=0.2) +
  geom_point(aes(ADM_CARRERA_NOMBRE, PORCENTAJE), show.legend = FALSE) + 
  coord_flip() + 
  labs(fill="Programa PAES") +
  ylab("Porcentaje de admitidos PAES") +
  xlab("Carreras") +
  ggtitle(label = "Admitidos PAES a la sede Palmira por carrera", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) +
  theme_classic() + 
  geom_hline(yintercept = 2, color="black") +
  facet_wrap(~PROGRAMA) 
```

### Estrato socio-económico de los PAES de poblaciones afrocolombianas

El **acuerdo 13 de 2009 del CSU en el articulo uno** para las _poblaciones afrocolombianas_ establece que los admitidos por esta modalidad deberan pertenecer a los estratos 1 ó 2. Se observa a continuacíon que, aunque la mayoría de los admitidos cumple con la normatividad, existen unos cuantos que o bien pertenecen al estrato 3 o bien se rotulan como estrato 0. Lo cual es claramente una falta de cumplimiento en a norma:

```{r,echo=FALSE}
df <-  PAES  %>%  group_by(ESTRATO_ORIG,PAES) %>% count() %>% replace_na(list(ESTRATO_ORIG="Sin información")) 

df2 <- df %>% ungroup() %>% group_by(PAES) %>% mutate("N"=sum(n)) %>% group_by(ESTRATO_ORIG) %>% mutate("Porcentaje"=round(n/N,2)*100) %>% filter(PAES=="Población afrocolombiana")

df <- df  %>% filter(PAES=="Población afrocolombiana")

color <- c( "#B23AEE")

ggplot(df, aes(ESTRATO_ORIG, n, fill=PAES)) + 
  geom_bar(stat = "identity", show.legend = FALSE)  + 
  ggtitle(label = "Admitidos PAES por estrato y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Sede de presencia nacional") +
  xlab("Estrato socioeconómico") +
  ylab("Número de estudiantes") + 
  theme_classic()  +
  facet_grid(.~ PAES) 

ggplot(df2, aes(ESTRATO_ORIG, Porcentaje, fill=PAES)) + 
  geom_bar(stat = "identity", show.legend = FALSE)  + 
  ggtitle(label = "Admitidos PAES por estrato y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Sede de presencia nacional") +
  xlab("Estrato socioeconómico") +
  ylab("Porcentaje de estudiantes") + 
  theme_classic() +
  facet_grid(.~ PAES) 
```



## Validación de la normatividad vigente y aplicable a los PEAMA

### Puntaje total de admisión

Como dice la **resolución 125 de 2008 de la Rectoría en el artículo cuatro** y la **resolución 1708 de 2009 de la Rectoría en el artículo dos**  para el programa _PEAMA Amazonía_, la **resolución 16 de 2008 de la rectoría en el artículo tres** para el programa _PEAMA Caribe_ y la **resolución 1302 de 2007 de la rectoría en el artículo cuatro** para el programa _PEAMA Orinoquía_, los estudiantes admitidos por estas modalidades deben tener un puntaje de admisión mayor o igual al último admitido de manera regular.

```{r,echo=FALSE}
df <- PEAMA %>% select(`SEDE PRESENCIA NACIONAL`,PUNTAJE_TOTAL)
df$PUNTAJE_TOTAL <- as.numeric(df$PUNTAJE_TOTAL)
color <- c("#A52A2A", "#CD6600", "#96CDCD")

ggplot(df, aes(`SEDE PRESENCIA NACIONAL`,PUNTAJE_TOTAL, fill=`SEDE PRESENCIA NACIONAL`)) +
  geom_boxplot(show.legend = FALSE) +
  scale_fill_manual(values = color) +
  xlab("Sede de presencia nacional") +
  ylab("Puntaje estandar de admisión") + 
  theme_classic() + 
  ggtitle(label = "Puntaje total de admisión de los admitidos PEAMA", subtitle =  "(Cohorte 2011-01)") +
  geom_hline(yintercept = as.numeric(PUNTAJE$P),color="red")
```

### Departamento de residencia de los admitidos

Como dice el **acuerdo 25 de 2007 del CSU en el artículo dos** los estudiantes que se presentan a través de la admisión en las sedes de presencia nacional deben tener domicilio en la región de alcance de la misma. la **resolución 125 de 2008 de la Rectoría en el artículo dos** establece para el programa _PEAMA Amazonía_ como zona de residencia los departamentos de Amazonía, Guainia, Putumayo o Vaupés, la **resolución 16 de 2008 de la rectoría en el artículo dos** establece para el programa _PEAMA Caribe_ que los aspirantes deben residir en el archipielago de San Andrés (mediante certificación de la OCCRE) y la **resolución 1302 de 2007 de la rectoría en el artículo dos** establece para el programa _PEAMA Orinoquía_ como zona de residencia los departamentos de Arauca, Casanare, Guainia, Guaviare y Vichada.


```{r,echo=FALSE}
df <- PEAMA_2 %>% group_by(DEP_RES,ADM_SEDE_NOMBRE) %>% summarise("ADMITIDOS"=n())

color <- c("#A52A2A", "#CD6600", "#96CDCD")
ggplot(df %>% filter(ADM_SEDE_NOMBRE=="Amazonía") , aes(DEP_RES, ADMITIDOS)) + 
  geom_bar(stat = "identity", position = "dodge", fill = color[1]) + 
  coord_flip() + 
  ggtitle(label = "Admitidos PEAMA - Amazonía por departamento de residencia", subtitle =  "(Cohorte 2011-01)") +
  xlab("Departamento") +
  ylab("Admitidos") + 
  theme_classic() 
```

Observamos que los admitidos a través del programa _PEAMA Amazonía_ provienen de los departamentos: Amazonía, Putumayo, Caquetá y Bogotá D.C., de donde se encuentra que los dos últimos departamentos se salen de la norma anteriormente descrita.

```{r,echo=FALSE}
color <- c("#A52A2A", "#CD6600", "#96CDCD")
ggplot(df %>% filter(ADM_SEDE_NOMBRE=="Caribe") , aes(DEP_RES, ADMITIDOS)) + 
  geom_bar(stat = "identity", position = "dodge", fill = color[2]) + 
  coord_flip() + 
  ggtitle(label = "Admitidos PEAMA - Caribe por departamento de residencia", subtitle =  "(Cohorte 2011-01)") +
  xlab("Departamento") +
  ylab("Admitidos") + 
  theme_classic() 
```

Para el caso del programa _PEAMA Caribe_ se encuentra que la residencia de los admitidos es de los departamentos: Archipielago de San Andrés, Providencia y Santa Catalina, Magdalena y Cesar, nuevamente estos dos últimos no concuerdan con la norma correspondiente.

```{r,echo=FALSE}
color <- c("#A52A2A", "#CD6600", "#96CDCD")
ggplot(df %>% filter(ADM_SEDE_NOMBRE=="Orinoquía") , aes(DEP_RES, ADMITIDOS)) + 
  geom_bar(stat = "identity", position = "dodge", fill = color[3]) + 
  coord_flip() + 
  ggtitle(label = "Admitidos PEAMA - Orinoquía por departamento de residencia", subtitle =  "(Cohorte 2011-01)") +
  xlab("Departamento") +
  ylab("Admitidos") + 
  theme_classic() 
```

Finalmente en el programa _PEAMA Orinoquía_ se encuentran que son residentes de los departamentos: Aráuca, Casanare, Guainia, Guaviare, Meta y Vichada. El departamento del Meta no se encuentra completado en la norma aplicable.

### Cupos por programa

Como se dispone en la **resolución 125 de 2008 de la Rectoría en el artículo tres** y ampliado en la **resolución 1708 de 2009 de la Rectoría en el artículo uno**  para el programa _PEAMA Amazonía_, la **resolución 16 de 2008 de la rectoría en el artículo tres** para el programa _PEAMA Caribe_ y la **resolución 1302 de 2007 de la rectoría en el artículo cuatro** para el programa _PEAMA Orinoquía_, los estudiantes admitidos por estas modalidades pueden tomar los cupos de admisión para las carreras establecidos en la norma y la cantidad de cupos tambien dispuestos ahí mismo. Se observa en color azúl los cupos por carrera dictados por la norma y en rojo los admitidos a los mismos.

```{r,echo=FALSE, warning=FALSE}
PEAMA[PEAMA$PROGRAMA=="Bogotá Contaduría Pública","PROGRAMA"] <- "Bogotá Contaduría"
PEAMA[PEAMA$PROGRAMA=="Bogotá Ciencia Política","PROGRAMA"] <- "Bogotá Ciencia política"
PEAMA[PEAMA$PROGRAMA=="Bogotá Ingeniería de Sistemas y Computación","PROGRAMA"] <- "Bogotá Ingeniería de Sistemas"
PEAMA[PEAMA$PROGRAMA=="Manizales Administración de Empresas (Diurno)","PROGRAMA"] <- "Manizales Administración de Empresas (diurna)"
PEAMA[PEAMA$PROGRAMA=="Manizales Gestión Cultural y Comunicativa","PROGRAMA"] <- "Manizales Gestión cultural"
PEAMA[PEAMA$PROGRAMA=="Medellín Ciencia Política","PROGRAMA"] <- "Medellín Ciencia política"

## Bogotá

P <- PEAMA %>% group_by(PROGRAMA,`SEDE PRESENCIA NACIONAL`) %>% summarise("ADMITIDOS"=n()) %>% ungroup() %>% distinct()
P <- P %>% left_join(CUPOS_PEAMA %>% select(PROGRAMA,SEDE), by=c("PROGRAMA"="PROGRAMA")) %>% distinct()
df <- P %>% left_join(CUPOS_PEAMA %>% select(-SEDE), by=c("PROGRAMA"="PROGRAMA","SEDE PRESENCIA NACIONAL"="SPN")) %>% filter(SEDE=="Bogotá")

df2 <- PEAMA %>% group_by(PROGRAMA,`SEDE PRESENCIA NACIONAL`) %>% summarise("ADMITIDOS"=n()) %>% ungroup() %>% anti_join(CUPOS_PEAMA,by=c("PROGRAMA"="PROGRAMA","SEDE PRESENCIA NACIONAL"="SPN"))
df <- df %>% gather("TIPO","CANTIDAD",CUPOS,ADMITIDOS)

color <- c("#B22222", "#79CDCD")
ggplot(df , aes(PROGRAMA, CANTIDAD, fill = TIPO)) + 
  geom_bar(stat = "identity", position = "dodge") + 
  scale_y_continuous(breaks=0:3,labels=0:3) + 
  coord_flip() +
  facet_grid(.~`SEDE PRESENCIA NACIONAL`)   + 
  ggtitle(label = "Cupos y admitidos PEAMA por programa en la Sede Bogotá por sede de presencia nacional", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill=" ") +
  xlab("Programa") +
  ylab("Cantidad") + 
  theme_classic() 

## Manizales

P <- PEAMA %>% group_by(PROGRAMA,`SEDE PRESENCIA NACIONAL`) %>% summarise("ADMITIDOS"=n()) %>% ungroup() %>% distinct()
P <- P %>% left_join(CUPOS_PEAMA %>% select(PROGRAMA,SEDE), by=c("PROGRAMA"="PROGRAMA")) %>% distinct()
df <- P %>% left_join(CUPOS_PEAMA %>% select(-SEDE), by=c("PROGRAMA"="PROGRAMA","SEDE PRESENCIA NACIONAL"="SPN")) %>% filter(SEDE=="Manizales")

df2 <- PEAMA %>% group_by(PROGRAMA,`SEDE PRESENCIA NACIONAL`) %>% summarise("ADMITIDOS"=n()) %>% ungroup() %>% anti_join(CUPOS_PEAMA,by=c("PROGRAMA"="PROGRAMA","SEDE PRESENCIA NACIONAL"="SPN"))
df <- df %>% gather("TIPO","CANTIDAD",CUPOS,ADMITIDOS)

color <- c("#B22222", "#79CDCD")
ggplot(df , aes(PROGRAMA, CANTIDAD, fill = TIPO)) + 
  geom_bar(stat = "identity", position = "dodge") + 
  scale_y_continuous(breaks=0:3,labels=0:3) + 
  coord_flip() +
  facet_grid(.~`SEDE PRESENCIA NACIONAL`) + 
  ggtitle(label = "Cupos y admitidos PEAMA por programa en la Sede Manizales por sede de presencia nacional", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill=" ") +
  xlab("Programa") +
  ylab("Cantidad") + 
  theme_classic() 

## Medellín

P <- PEAMA %>% group_by(PROGRAMA,`SEDE PRESENCIA NACIONAL`) %>% summarise("ADMITIDOS"=n()) %>% ungroup() %>% distinct()
P <- P %>% left_join(CUPOS_PEAMA %>% select(PROGRAMA,SEDE), by=c("PROGRAMA"="PROGRAMA")) %>% distinct()
df <- P %>% left_join(CUPOS_PEAMA %>% select(-SEDE), by=c("PROGRAMA"="PROGRAMA","SEDE PRESENCIA NACIONAL"="SPN")) %>% filter(SEDE=="Medellín")

df2 <- PEAMA %>% group_by(PROGRAMA,`SEDE PRESENCIA NACIONAL`) %>% summarise("ADMITIDOS"=n()) %>% ungroup() %>% anti_join(CUPOS_PEAMA,by=c("PROGRAMA"="PROGRAMA","SEDE PRESENCIA NACIONAL"="SPN"))
df <- df %>% gather("TIPO","CANTIDAD",CUPOS,ADMITIDOS)

color <- c("#B22222", "#79CDCD")
ggplot(df , aes(PROGRAMA, CANTIDAD, fill = TIPO)) + 
  geom_bar(stat = "identity", position = "dodge")+ 
  scale_y_continuous(breaks=0:3,labels=0:3)  + 
  coord_flip() +
  facet_grid(.~`SEDE PRESENCIA NACIONAL`)  + 
  ggtitle(label = "Cupos y admitidos PEAMA por programa en la Sede Medellín por sede de presencia nacional", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill=" ") +
  xlab("Programa") +
  ylab("Cantidad") + 
  theme_classic() 

## Palmira

P <- PEAMA %>% group_by(PROGRAMA,`SEDE PRESENCIA NACIONAL`) %>% summarise("ADMITIDOS"=n()) %>% ungroup() %>% distinct()
P <- P %>% left_join(CUPOS_PEAMA %>% select(PROGRAMA,SEDE), by=c("PROGRAMA"="PROGRAMA")) %>% distinct()
df <- P %>% left_join(CUPOS_PEAMA %>% select(-SEDE), by=c("PROGRAMA"="PROGRAMA","SEDE PRESENCIA NACIONAL"="SPN")) %>% filter(SEDE=="Palmira")

df2 <- PEAMA %>% group_by(PROGRAMA,`SEDE PRESENCIA NACIONAL`) %>% summarise("ADMITIDOS"=n()) %>% ungroup() %>% anti_join(CUPOS_PEAMA,by=c("PROGRAMA"="PROGRAMA","SEDE PRESENCIA NACIONAL"="SPN"))
df <- df %>% gather("TIPO","CANTIDAD",CUPOS,ADMITIDOS)

color <- c("#B22222", "#79CDCD")
ggplot(df , aes(PROGRAMA, CANTIDAD, fill = TIPO)) + 
  geom_bar(stat = "identity", position = "dodge")+ 
  scale_y_continuous(breaks=0:3,labels=0:3)  + 
  coord_flip() +
  facet_grid(.~`SEDE PRESENCIA NACIONAL`)   + 
  ggtitle(label = "Cupos y admitidos PEAMA por programa en la Sede Palmira por sede de presencia nacional", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill=" ") +
  xlab("Programa") +
  ylab("Cantidad") + 
  theme_classic() 

```

Se observa que la norma sólo está siendo violada por seis admitidos del _PEAMA AMAZONÍA_ a carreras de Medellín que no se encontraban aprobadas por las normas, específicamente dos a Ingeniería Administrativa, uno a Ingeniería de Minas y Metalurgia, dos a Ingeniería de Petroleos y uno a Zootecnia.

### Total de cupos

De acuerdo al **acuerdo 25 de 2007 del CSU en el considerando cuatro** se espera la universidad otorgue el $1\%$ de cupos a bachilleres de los departamentos en los que no haya instituciones de educación superior y otro $1\%$ a los aspirantes que provengan de municipios de difícil acceso o con problemas de orden público, dando así una norma de $2\%$ de cupos que deben ser otorgados por un sistema especial de admisión en la Universidad. El programa PEAMA, de acuerdo a las normas **resolución 125 de 2008 de la Rectoría en el artículo tres** y ampliado en la **resolución 1708 de 2009 de la Rectoría en el artículo uno**  para el programa _PEAMA Amazonía_, la **resolución 16 de 2008 de la rectoría en el artículo tres** para el programa _PEAMA Caribe_ y la **resolución 1302 de 2007 de la rectoría en el artículo cuatro** para el programa _PEAMA Orinoquía_, establece un total de 190 cupos para estos programas, dado que para a cohorte de estudio el número total de cupos disponibles para admisión es del 5083 cupos según la **resolución que la DNA nos debería indicar**. De donde el porcentaje de cupos de admisión otorgados por esta modalidad especial es del $3.8\%$, dando así un cumplimiento satisfactorio de la norma.
