==> 08-permanenciaPAES.Rmd <==
# Permanencia PAES

```{r,eval=TRUE,echo=FALSE,results='hide',warning=FALSE,error=FALSE,message=FALSE}
library("tidyverse")
load("PAES.RData")
PAES[PAES$SUBACCESO=="INDÍGENA","SUBACCESO"] <- "Comunidades indígenas"
PAES[PAES$SUBACCESO=="MEJOR BACHILLER","SUBACCESO"] <- "Mejores bachilleres"
PAES[PAES$SUBACCESO=="MUNICIPIO","SUBACCESO"] <- "Mejores bachilleres de municipios pobres"
PAES[PAES$SUBACCESO=="POBLACION NEGRA, AFROCOLOMBIANA, PALENQUERA Y RAIZAL","SUBACCESO"] <- "Población afrocolombiana"
```

## Máximo périodo de avance 


```{r,echo=FALSE}
df <- PAES %>%  group_by(`PERIODO AVANCE`,SUBACCESO) %>% count()

df2 <- df %>% ungroup() %>% group_by(SUBACCESO) %>% mutate("N"=sum(n)) %>% group_by(`PERIODO AVANCE`) %>% mutate("Porcentaje"=round(n/N,2)*100)


color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df, aes(`PERIODO AVANCE`, n, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Matriculados PAES según máximo periodo de avance y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programa") +
  xlab("Máximo periodo de avance") +
  ylab("Número de estudiantes") + 
  theme_classic() + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(df) + 
  geom_line(aes(`PERIODO AVANCE`, n, color=SUBACCESO,group = SUBACCESO), show.legend = FALSE)  + 
  geom_point(aes(`PERIODO AVANCE`, n, color=SUBACCESO), show.legend = FALSE)  + 
  ggtitle(label = "Matriculados PAES según máximo periodo de avance y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) + 
  labs(color="Sede de presencia nacional") +
  xlab("Máximo periodo de avance") +
  ylab("Número de estudiantes") + 
  theme_classic()  +
  facet_wrap(~SUBACCESO) + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(df2) + 
  geom_line(aes(`PERIODO AVANCE`, Porcentaje, color=SUBACCESO,group = SUBACCESO))  + 
  geom_point(aes(`PERIODO AVANCE`, Porcentaje, color=SUBACCESO))  + 
  ggtitle(label = "Matriculados PAES según máximo periodo de avance y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) + 
  labs(color="Sede de presencia nacional") +
  xlab("Máximo periodo de avance") +
  ylab("Pordentaje de de estudiantes dentro de cada grupo") + 
  theme_classic() + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))
```



## Matriculados por semestre


```{r,echo=FALSE}
df <- PAES %>% select(SUBACCESO,`2011-01`,`2011-03`,`2012-01`,`2012-03`,`2013-01`,`2013-03`,`2014-01`,`2014-03`,`2015-01`,`2015-03`,`2016-01`,`2016-03`) %>% gather(key = "YEAR", value = "SI",`2011-01`:`2016-03`,na.rm = TRUE) %>%  group_by(YEAR, SUBACCESO) %>%  count() %>% arrange(YEAR)


df2 <- df %>% ungroup() %>% group_by(SUBACCESO) %>% mutate("N"=sum(n)) %>% group_by(YEAR) %>% mutate("Porcentaje"=round(n/N,2)*100)


color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df, aes(YEAR, n, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Matriculados PAES por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programa") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic() + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))


ggplot(df) + 
  geom_line(aes(YEAR, n, color=SUBACCESO,group = SUBACCESO))  + 
  geom_point(aes(YEAR, n, color=SUBACCESO))  + 
  ggtitle(label = "Matriculados PAES por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) + 
  labs(color="Sede de presencia nacional") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic() + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(df2) + 
  geom_line(aes(YEAR, Porcentaje, color=SUBACCESO,group = SUBACCESO), show.legend = FALSE)  + 
  geom_point(aes(YEAR, Porcentaje, color=SUBACCESO), show.legend = FALSE)  + 
  ggtitle(label = "Matriculados PAES por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) + 
  labs(color="Sede de presencia nacional") +
  xlab("Periodo") +
  ylab("Porcentaje de estudiantes") + 
  theme_classic() +
  facet_wrap(~SUBACCESO) + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))
```


## Cancelaciones por semestre


```{r,echo=FALSE}
df <- PAES %>% select(SUBACCESO, `2011-1 Cancelacion SA`,`2011-3 Cancelacion SA`,`2012-1 Cancelacion SA`,`2012-3 Cancelacion SA`,`2013-1 Cancelacion SA`,`2013-3 Cancelacion SA`,`2014-1 Cancelacion SA`,`2014-3 Cancelacion SA`,`2015-1 Cancelacion SA`,`2015-3 Cancelacion SA`,`2016-1 Cancelacion SA`,`2016-3 Cancelacion SA`) %>% gather(key = "YEAR", value = "SI",`2011-1 Cancelacion SA`:`2016-3 Cancelacion SA`,na.rm = TRUE) %>%  group_by(YEAR, SUBACCESO) %>%  summarise(n=sum(SI)) %>% arrange(YEAR) %>% separate(YEAR,c("YEAR","T1","T2"), sep = " ") %>%  select(-T1,-T2)

df2 <- df %>% ungroup() %>% group_by(SUBACCESO) %>% mutate("N"=sum(n)) %>% group_by(YEAR) %>% mutate("Porcentaje"=round(n/N,2)*100)


color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df, aes(YEAR, n, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Estudiantes PAES que cancelaron por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programa") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic()+ 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(df) + 
  geom_line(aes(YEAR, n, color=SUBACCESO,group = SUBACCESO), show.legend = FALSE)  + 
  geom_point(aes(YEAR, n, color=SUBACCESO), show.legend = FALSE)  + 
  ggtitle(label = "Estudiantes PAES que cancelaron por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) + 
  labs(color="Sede de presencia nacional") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic()  + 
  theme_classic() +
  facet_wrap(~SUBACCESO) + 
  scale_y_continuous(breaks=2*(1:5),labels=2*(1:5)) + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(df2) + 
  geom_line(aes(YEAR, Porcentaje, color=SUBACCESO,group = SUBACCESO), show.legend = FALSE)  + 
  geom_point(aes(YEAR, Porcentaje, color=SUBACCESO), show.legend = FALSE)  + 
  ggtitle(label = "Estudiantes PAES que cancelaron por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) + 
  labs(color="Sede de presencia nacional") +
  xlab("Periodo") +
  ylab("Porcentaje de estudiantes") + 
  theme_classic() +
  facet_wrap(~SUBACCESO) + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))
```


## Deserciones por semestre

```{r,echo=FALSE}
df <- PAES %>% filter(DESENLACE=="dac"|DESENLACE=="dnac") %>% group_by(SUBACCESO, PERIODO_DESENLACE) %>% count()

df <- df %>% filter(PERIODO_DESENLACE!="2017-03")

df2 <- df %>% ungroup() %>% group_by(SUBACCESO) %>% mutate("N"=sum(n)) %>% group_by(PERIODO_DESENLACE) %>% mutate("Porcentaje"=round(n/N,2)*100)


color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df, aes(PERIODO_DESENLACE, n, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Estudiantes PAES que desertaron por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programa") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic()+ 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))


ggplot(df) + 
  geom_line(aes(PERIODO_DESENLACE, n, color=SUBACCESO,group = SUBACCESO), show.legend = FALSE)  + 
  geom_point(aes(PERIODO_DESENLACE, n, color=SUBACCESO), show.legend = FALSE)  + 
  ggtitle(label = "Estudiantes PAES que desertaron por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) + 
  labs(color="Sede de presencia nacional") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic()  + 
  theme_classic() +
  facet_wrap(~SUBACCESO) + 
  #scale_y_continuous(breaks=2*(1:5),labels=2*(1:5)) + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(df2) + 
  geom_line(aes(PERIODO_DESENLACE, Porcentaje, color=SUBACCESO,group = SUBACCESO), show.legend = FALSE)  + 
  geom_point(aes(PERIODO_DESENLACE, Porcentaje, color=SUBACCESO), show.legend = FALSE)  + 
  ggtitle(label = "Estudiantes PAES que desertaron por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) + 
  labs(color="Sede de presencia nacional") +
  xlab("Periodo") +
  ylab("Porcentaje de estudiantes") + 
  theme_classic() +
  facet_wrap(~SUBACCESO) + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

```

## Deserciones académicas por semestre


```{r,echo=FALSE}
df <- PAES %>% filter(DESENLACE=="dac") %>% group_by(SUBACCESO, PERIODO_DESENLACE) %>% count()

df2 <- df %>% ungroup() %>% group_by(SUBACCESO) %>% mutate("N"=sum(n)) %>% group_by(PERIODO_DESENLACE) %>% mutate("Porcentaje"=round(n/N,2)*100)

color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df, aes(PERIODO_DESENLACE, n, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Estudiantes PAES que desertaron debido a razones académicas por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programa") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic()+ 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(df) + 
  geom_line(aes(PERIODO_DESENLACE, n, color=SUBACCESO,group = SUBACCESO), show.legend = FALSE)  + 
  geom_point(aes(PERIODO_DESENLACE, n, color=SUBACCESO), show.legend = FALSE)  + 
  ggtitle(label = "Estudiantes PAES que desertaron debido a razones académicas por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) + 
  labs(color="Sede de presencia nacional") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic()  + 
  theme_classic() +
  facet_wrap(~SUBACCESO) + 
  #scale_y_continuous(breaks=2*(1:5),labels=2*(1:5)) + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(df2) + 
  geom_line(aes(PERIODO_DESENLACE, Porcentaje, color=SUBACCESO,group = SUBACCESO), show.legend = FALSE)  + 
  geom_point(aes(PERIODO_DESENLACE, Porcentaje, color=SUBACCESO), show.legend = FALSE)  + 
  ggtitle(label = "Estudiantes PAES que desertaron debido a razones académicas por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) + 
  labs(color="Sede de presencia nacional") +
  xlab("Periodo") +
  ylab("Porcentaje de estudiantes") + 
  theme_classic() +
  facet_wrap(~SUBACCESO) + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))
```

## Deserciones no académicas por semestre


```{r,echo=FALSE}
df <- PAES %>% filter(DESENLACE=="dnac") %>% group_by(SUBACCESO, PERIODO_DESENLACE) %>% count()

df <- df %>% filter(PERIODO_DESENLACE!="2017-03")

df2 <- df %>% ungroup() %>% group_by(SUBACCESO) %>% mutate("N"=sum(n)) %>% group_by(PERIODO_DESENLACE) %>% mutate("Porcentaje"=round(n/N,2)*100)

color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df, aes(PERIODO_DESENLACE, n, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Estudiantes PAES que desertaron debido a razones no académicas por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programa") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic()+ 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(df) + 
  geom_line(aes(PERIODO_DESENLACE, n, color=SUBACCESO,group = SUBACCESO), show.legend = FALSE)  + 
  geom_point(aes(PERIODO_DESENLACE, n, color=SUBACCESO), show.legend = FALSE)  + 
  ggtitle(label = "Estudiantes PAES que desertaron debido a razones no académicas por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) + 
  labs(color="Sede de presencia nacional") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic()  + 
  theme_classic() +
  facet_wrap(~SUBACCESO) + 
  #scale_y_continuous(breaks=2*(1:5),labels=2*(1:5)) + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(df2) + 
  geom_line(aes(PERIODO_DESENLACE, Porcentaje, color=SUBACCESO,group = SUBACCESO), show.legend = FALSE)  + 
  geom_point(aes(PERIODO_DESENLACE, Porcentaje, color=SUBACCESO), show.legend = FALSE)  + 
  ggtitle(label = "Estudiantes PAES que desertaron debido a razones no académicas por periodo y programa", subtitle =  "(Cohorte 2011-01)") +
  scale_color_manual(values = color) + 
  labs(color="Sede de presencia nacional") +
  xlab("Periodo") +
  ylab("Porcentaje de estudiantes") + 
  theme_classic() +
  facet_wrap(~SUBACCESO) + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1))
```

## Promedio por semestre


```{r,echo=FALSE, warning=FALSE}
df <- PAES %>% select(SUBACCESO, `Promedio 2011-1SA`,`Promedio 2011-3SA`,`Promedio 2012-1SA`,`Promedio 2012-3SA`,`Promedio 2013-1SA`,`Promedio 2013-3SA`,`Promedio 2014-1SA`,`Promedio 2014-3SA`,`Promedio 2015-1SA`,`Promedio 2015-3SA`,`Promedio 2016-1SA`,`Promedio 2016-3SA`) %>% gather(key = "YEAR", value = "Promedio",`Promedio 2011-1SA`:`Promedio 2016-3SA`,na.rm = TRUE)   %>% separate(YEAR,c("T1","YEAR"), sep = " ") %>%   select(-T1) %>% separate(YEAR,c("YEAR","T1"), sep = "SA") %>% select(-T1) %>% arrange(YEAR)


color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df, aes(YEAR,Promedio, fill=SUBACCESO)) +
  geom_boxplot() +
  scale_fill_manual(values = color) +
  xlab("Periodo") +
  ylab("Promedio") + 
  theme_classic() + 
  ggtitle(label = "Promedio en el semestre de los matriculados PAES", subtitle =  "(Cohorte 2011-01)")+
  coord_flip()

ggplot(df, aes(YEAR,Promedio, fill=SUBACCESO)) +
  geom_boxplot(outlier.shape = NA) +
  scale_y_continuous(limits = c(1.5, 5)) +
  scale_fill_manual(values = color) +
  xlab("Periodo") +
  ylab("Promedio") + 
  theme_classic() + 
  ggtitle(label = "Promedio en el semestre de los matriculados PAES", subtitle =  "(Cohorte 2011-01)")+ coord_flip()

ggplot(df, aes(YEAR,Promedio, fill=SUBACCESO)) +
  geom_boxplot( show.legend = FALSE) +
  scale_fill_manual(values = color) +
  xlab("Periodo") +
  ylab("Promedio") + 
  theme_classic() + 
  ggtitle(label = "Promedio en el semestre de los matriculados PAES", subtitle =  "(Cohorte 2011-01)")+ 
  coord_flip()+
  facet_wrap(~SUBACCESO)

ggplot(df, aes(YEAR,Promedio, fill=SUBACCESO)) +
  geom_boxplot(outlier.shape = NA, show.legend = FALSE) +
  scale_y_continuous(limits = c(1.5, 5)) +
  scale_fill_manual(values = color) +
  xlab("Periodo") +
  ylab("Promedio") + 
  theme_classic() + 
  ggtitle(label = "Promedio en el semestre de los matriculados PAES", subtitle =  "(Cohorte 2011-01)")+ 
  coord_flip() +
  facet_wrap(~SUBACCESO)
```


## PAPA por semestre


```{r,echo=FALSE, warning=FALSE}
df <- PAES %>% select(SUBACCESO, `PAPA 2011-1SA`,`PAPA 2011-3SA`,`PAPA 2012-1SA`,`PAPA 2012-3SA`,`PAPA 2013-1SA`,`PAPA 2013-3SA`,`PAPA 2014-1SA`,`PAPA 2014-3SA`,`PAPA 2015-1SA`,`PAPA 2015-3SA`,`PAPA 2016-1SA`,`PAPA 2016-3SA`) %>% gather(key = "YEAR", value = "PAPA",`PAPA 2011-1SA`:`PAPA 2016-3SA`,na.rm = TRUE)   %>% separate(YEAR,c("T1","YEAR"), sep = " ") %>%   select(-T1) %>% separate(YEAR,c("YEAR","T1"), sep = "SA") %>% select(-T1) %>% arrange(YEAR)

color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df, aes(YEAR,PAPA, fill=SUBACCESO)) +
  geom_boxplot() +
  scale_fill_manual(values = color) +
  xlab("Periodo") +
  ylab("PAPA") + 
  theme_classic() + 
  ggtitle(label = "PAPA por semestre de los matriculados PAES", subtitle =  "(Cohorte 2011-01)")+
  coord_flip()

ggplot(df, aes(YEAR,PAPA, fill=SUBACCESO)) +
  geom_boxplot(outlier.shape = NA) +
  scale_y_continuous(limits = c(1.5, 5)) +
  scale_fill_manual(values = color) +
  xlab("Periodo") +
  ylab("PAPA") + 
  theme_classic() + 
  ggtitle(label = "PAPA por semestre de los matriculados PAES", subtitle =  "(Cohorte 2011-01)")+ coord_flip()

ggplot(df, aes(YEAR,PAPA, fill=SUBACCESO)) +
  geom_boxplot( show.legend = FALSE) +
  scale_fill_manual(values = color) +
  xlab("Periodo") +
  ylab("PAPA") + 
  theme_classic() + 
  ggtitle(label = "PAPA por semestre de los matriculados PAES", subtitle =  "(Cohorte 2011-01)")+ 
  coord_flip()+
  facet_wrap(~SUBACCESO)

ggplot(df, aes(YEAR,PAPA, fill=SUBACCESO)) +
  geom_boxplot(outlier.shape = NA, show.legend = FALSE) +
  scale_y_continuous(limits = c(1.5, 5)) +
  scale_fill_manual(values = color) +
  xlab("Periodo") +
  ylab("PAPA") + 
  theme_classic() + 
  ggtitle(label = "PAPA por semestre de los matriculados PAES", subtitle =  "(Cohorte 2011-01)")+ 
  coord_flip() +
  facet_wrap(~SUBACCESO)
```

## Traslados de carrera

```{r,echo=FALSE}
df <- PAES %>%  filter(Traslado==1) %>% group_by(SUBACCESO, `PERIODO desde el que inició el cambio`) %>% count()

color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df, aes(`PERIODO desde el que inició el cambio`, n, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Cambios de carrera en los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic() 

ggplot(df, aes(`PERIODO desde el que inició el cambio`, n, fill=SUBACCESO)) + 
  geom_bar(stat = "identity", show.legend = FALSE)  + 
  ggtitle(label = "Cambios de carrera en los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic()  +
  facet_wrap(~SUBACCESO)  + 
  scale_y_continuous(breaks=1:2,labels=1:2) 
```

## Traslados de sede

```{r,echo=FALSE}
df <- PAES %>%  filter(`Traslado entre sedes`==1) %>% group_by(SUBACCESO, `PERIODO desde el que inició el cambio`) %>% count()

color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df, aes(`PERIODO desde el que inició el cambio`, n, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Cambios de sede en los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic() + 
  scale_y_continuous(breaks=1:2,labels=1:2) 

ggplot(df, aes(`PERIODO desde el que inició el cambio`, n, fill=SUBACCESO)) + 
  geom_bar(stat = "identity", show.legend = FALSE)  + 
  ggtitle(label = "Cambios de sede en los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("Periodo") +
  ylab("Número de estudiantes") + 
  theme_classic()  +
  facet_wrap(~SUBACCESO)  + 
  scale_y_continuous(breaks=1:2,labels=1:2) 

```

## Completó ciclo de estudios previsto por periodo

```{r,echo=FALSE,message=FALSE}
df <- PAES %>%  filter(DESENLACE=="grad") %>% group_by(SUBACCESO, PERIODO_DESENLACE) %>% count()

df <- df %>% filter(PERIODO_DESENLACE!="2017-03" & PERIODO_DESENLACE!="2017-01")

df2 <- df %>% ungroup() %>%  left_join(PAES %>% group_by(SUBACCESO) %>% mutate("N"=n()) %>%  select(SUBACCESO,N) %>%  distinct()) %>% mutate("Porcentaje"=round(n/N,2)*100)

color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df, aes(PERIODO_DESENLACE, n, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Completó ciclo de estudios previsto de los matriculados PAES por programa y periodo", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("Periodo") +
  ylab("Completaron ciclo de estudios previsto") + 
  theme_classic() 

ggplot(df, aes(PERIODO_DESENLACE, n, fill=SUBACCESO)) + 
  geom_bar(stat = "identity", show.legend = FALSE)  + 
  ggtitle(label = "Completó ciclo de estudios previsto de los matriculados PAES por programa y periodo", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("Periodo") +
  ylab("Completaron ciclo de estudios previsto") + 
  theme_classic()  +
  facet_wrap(~SUBACCESO)

ggplot(df2, aes(PERIODO_DESENLACE, Porcentaje, fill=SUBACCESO)) + 
  geom_bar(stat = "identity", show.legend = FALSE)  + 
  ggtitle(label = "Completó ciclo de estudios previsto de los matriculados PAES por programa y periodo", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("Periodo") +
  ylab("Porcentaje que completó ciclo de estudios previsto") + 
  theme_classic()  +
  facet_wrap(~SUBACCESO)
```

## Completó ciclo de estudios previsto total

```{r,echo=FALSE,message=FALSE}
df <- PAES  %>% filter(PERIODO_DESENLACE!="2017-03" & PERIODO_DESENLACE!="2017-01") %>%  filter(DESENLACE=="grad") %>% group_by(SUBACCESO) %>% count()

df2 <- df %>% ungroup() %>%  left_join(PAES %>% group_by(SUBACCESO) %>% mutate("N"=n()) %>%  select(SUBACCESO,N) %>%  distinct()) %>% mutate("Porcentaje"=round(n/N,2)*100)

df3 <- df2 %>% rename("Graduados"=n,"Matriculados"=N) %>% gather("CATEGORIA","TOTAL",Graduados,Matriculados,Porcentaje)

color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df3 %>%  filter(CATEGORIA!="Porcentaje"), aes(CATEGORIA, TOTAL, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Completó ciclo de estudios previsto de los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("") +
  ylab("") + 
  theme_classic() 

ggplot(df3%>%  filter(CATEGORIA!="Porcentaje"), aes(CATEGORIA, TOTAL, fill=SUBACCESO)) + 
  geom_bar(stat = "identity", show.legend = FALSE)  + 
  ggtitle(label = "Completó ciclo de estudios previsto de los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("") +
  ylab("") + 
  theme_classic()  +
  facet_wrap(~SUBACCESO)

ggplot(df2, aes(SUBACCESO, Porcentaje, fill=SUBACCESO)) + 
  geom_bar(stat = "identity", position="dodge", show.legend = FALSE)  + 
  ggtitle(label = "Completó ciclo de estudios previsto de los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("") +
  ylab("Porcentaje de estudiantes que completó ciclo de estudios previsto") + 
  theme_classic() +
  coord_flip()
```


## Desertores total

```{r,echo=FALSE,message=FALSE}
df <- PAES  %>% filter(PERIODO_DESENLACE!="2017-03" & PERIODO_DESENLACE!="2017-01") %>%  filter(DESENLACE=="dac"|DESENLACE=="dnac") %>% group_by(SUBACCESO) %>% count()

df2 <- df %>% ungroup() %>%  left_join(PAES %>% group_by(SUBACCESO) %>% mutate("N"=n()) %>%  select(SUBACCESO,N) %>%  distinct()) %>% mutate("Porcentaje"=round(n/N,2)*100)

df3 <- df2 %>% rename("Desertores"=n,"Matriculados"=N) %>% gather("CATEGORIA","TOTAL",Desertores,Matriculados,Porcentaje)

color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df3 %>%  filter(CATEGORIA!="Porcentaje"), aes(CATEGORIA, TOTAL, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Desertores en los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("") +
  ylab("") + 
  theme_classic() 

ggplot(df3 %>%  filter(CATEGORIA!="Porcentaje"), aes(CATEGORIA, TOTAL, fill=SUBACCESO)) + 
  geom_bar(stat = "identity", show.legend = FALSE)  + 
  ggtitle(label = "Desertores en los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("") +
  ylab("") + 
  theme_classic()  +
  facet_wrap(~SUBACCESO)

ggplot(df2, aes(SUBACCESO, Porcentaje, fill=SUBACCESO)) + 
  geom_bar(stat = "identity", position="dodge", show.legend = FALSE)  + 
  ggtitle(label = "Desertores en los\nmatriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("") +
  ylab("Porcentaje de estudiantes bloqueados") + 
  theme_classic() +
  coord_flip()
```

### Desertores académicos total

```{r,echo=FALSE,message=FALSE}
df <- PAES  %>% filter(PERIODO_DESENLACE!="2017-03" & PERIODO_DESENLACE!="2017-01") %>%  filter(DESENLACE=="dac") %>% group_by(SUBACCESO) %>% count()

df2 <- df %>% ungroup() %>%  left_join(PAES %>% group_by(SUBACCESO) %>% mutate("N"=n()) %>%  select(SUBACCESO,N) %>%  distinct()) %>% mutate("Porcentaje"=round(n/N,2)*100)

df3 <- df2 %>% rename("Desertores"=n,"Matriculados"=N) %>% gather("CATEGORIA","TOTAL",Desertores,Matriculados,Porcentaje)

color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df3 %>%  filter(CATEGORIA!="Porcentaje"), aes(CATEGORIA, TOTAL, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Desertores académicos en los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("") +
  ylab("") + 
  theme_classic() 

ggplot(df3 %>%  filter(CATEGORIA!="Porcentaje"), aes(CATEGORIA, TOTAL, fill=SUBACCESO)) + 
  geom_bar(stat = "identity", show.legend = FALSE)  + 
  ggtitle(label = "Desertores académicos en los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("") +
  ylab("") + 
  theme_classic()  +
  facet_wrap(~SUBACCESO)

ggplot(df2, aes(SUBACCESO, Porcentaje, fill=SUBACCESO)) + 
  geom_bar(stat = "identity", position="dodge", show.legend = FALSE)  + 
  ggtitle(label = "Desertores académicos en los\nmatriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("") +
  ylab("Porcentaje de estudiantes bloqueados") + 
  theme_classic() +
  coord_flip()
```

### Desertores no académicos total

```{r,echo=FALSE,message=FALSE}
df <- PAES %>% filter(PERIODO_DESENLACE!="2017-03" & PERIODO_DESENLACE!="2017-01") %>%  filter(DESENLACE=="dnac") %>% group_by(SUBACCESO) %>% count()

df2 <- df %>% ungroup() %>%  left_join(PAES %>% group_by(SUBACCESO) %>% mutate("N"=n()) %>%  select(SUBACCESO,N) %>%  distinct()) %>% mutate("Porcentaje"=round(n/N,2)*100)

df3 <- df2 %>% rename("Desertores"=n,"Matriculados"=N) %>% gather("CATEGORIA","TOTAL",Desertores,Matriculados,Porcentaje)

color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df3 %>%  filter(CATEGORIA!="Porcentaje"), aes(CATEGORIA, TOTAL, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Desertores no académicos en los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("") +
  ylab("") + 
  theme_classic() 

ggplot(df3 %>%  filter(CATEGORIA!="Porcentaje"), aes(CATEGORIA, TOTAL, fill=SUBACCESO)) + 
  geom_bar(stat = "identity", show.legend = FALSE)  + 
  ggtitle(label = "Desertores no académicos en los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("") +
  ylab("") + 
  theme_classic()  +
  facet_wrap(~SUBACCESO)

ggplot(df2, aes(SUBACCESO, Porcentaje, fill=SUBACCESO)) + 
  geom_bar(stat = "identity", position="dodge", show.legend = FALSE)  + 
  ggtitle(label = "Desertores no académicos en los\nmatriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("") +
  ylab("Porcentaje de estudiantes bloqueados") + 
  theme_classic() +
  coord_flip()
```

## Doble titulación total - no hay

```{r,echo=FALSE,message=FALSE}
df <- PAES %>%  filter(`Bloqueo doble titulación SA`==1) %>% group_by(SUBACCESO) %>% count()

df2 <- df %>% ungroup() %>%  left_join(PAES %>% group_by(SUBACCESO) %>% mutate("N"=n()) %>%  select(SUBACCESO,N) %>%  distinct()) %>% mutate("Porcentaje"=round(n/N,2))

df3 <- df2 %>% rename("Bloqueados"=n,"Matriculados"=N) %>% gather("CATEGORIA","TOTAL",Bloqueados,Matriculados,Porcentaje)

color <- c("#1E90FF", "#CD0000", "#6E8B3D", "#B23AEE")

ggplot(df3 %>%  filter(CATEGORIA!="Porcentaje"), aes(CATEGORIA, TOTAL, fill=SUBACCESO)) + 
  geom_bar(stat = "identity")  + 
  ggtitle(label = "Bloqueo por doble titulación en los matriculados PAES por programa", subtitle =  "(Cohorte 2011-01)") +
  scale_fill_manual(values = color) + 
  labs(fill="Programas PAES") +
  xlab("Periodo") +
  ylab("") + 
  theme_classic() 
```
